automata simplification and alpha version
Big changes: decided to merge what ambiguous, non-ambiguous and parameters states and rather check at run time in the future the ambiguosity. It simplifies a lot the automata structure and makes many things much simpler, whereas it makes sense that the necessary checks and particular implementations of the group sign or other grammar modifers are made with dedicated functions at run time rather than inside the automata itself.

so we have an alpha version of the automata, without omissions yet.